requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
uuid==1.30
Pillow==5.2.0

